requirements.txt
### Base ###
django
djangorestframework

### Celery beat ###
django-celery-beat

### Cache ###
django-redis

### DataBase ###
mysqlclient

### Metrics ###
django-prometheus

### Versioning ###
bumpversion==0.6.0